+++ src/app/eth-extended/module/account/balance/component/ChartSection.tsx
@@ -12,6 +12,7 @@ import { IPortfolioChartData, CHART_DATA_KEY } from "./chart/BalanceChartData";
 import { IBalanceAreaChartPayload } from "./chart/IBalanceAreaChartPayload";
 import { BALANCE_HISTORY_DAYS } from "app/eth-extended/data/account/balance/AccountBalanceStore";
 import { IAsyncData } from "plugin-api/IAsyncData";
+import { weiToEth } from "app/util/wei";
 
 export interface IChartSectionProps {
     accountBalance: IAsyncData<AccountBalance>;
@@ -19,12 +20,17 @@ export interface IChartSectionProps {
     translation: ITranslation;
     locale: string;
     ethSymbol: string;
+    usdPricesEnabled: boolean;
+    /** Which token to show in the chart (or ETH if unset), without 0x prefix */
+    tokenAddress?: string;
 }
 
 @observer
 export class ChartSection extends React.Component<IChartSectionProps> {
     render() {
-        let { accountBalance, isFreshAccount, translation: tr, locale, ethSymbol } = this.props;
+        let {
+            accountBalance, isFreshAccount, translation: tr, locale, ethSymbol, usdPricesEnabled
+        } = this.props;
 
         return (
             <ChartContainer>
@@ -37,7 +43,10 @@ export class ChartSection extends React.Component<IChartSectionProps> {
                     data={!isFreshAccount && accountBalance.isLoaded() ?
                         this.computeChartData(accountBalance.data) : this.getPlaceholderChartData()}
                     locale={locale}
-                    ethSymbol={ethSymbol}
+                    ethSymbol={this.props.tokenAddress && accountBalance.isLoaded() ?
+                        accountBalance.data.getTokenBalance(this.props.tokenAddress).currency.symbol :
+                        ethSymbol}
+                    usdPricesEnabled={usdPricesEnabled}
                     disabled={isFreshAccount || !accountBalance.isLoaded()}
                 />
             </ChartContainer>
@@ -50,16 +59,53 @@ export class ChartSection extends React.Component<IChartSectionProps> {
             min: +Infinity,
             points: []
         };
-        data.points = accountBalance.computeTotalBalance().map(balance => {
-            let point: IBalanceAreaChartPayload = {
-                [CHART_DATA_KEY]: balance.usd,
-                balanceWei: balance.wei,
-                timestamp: balance.timestamp
-            };
-            data.min = Math.min(data.min, balance.usd);
-            data.max = Math.max(data.max, balance.usd);
-            return point;
-        }).reverse();
+
+        let { tokenAddress, usdPricesEnabled } = this.props;
+
+        if (tokenAddress) {
+                let balanceData = accountBalance.getTokenBalance(tokenAddress);
+
+                data.points = balanceData.chart.map(balance => {
+                    let tokenVolume = balance.balance.shiftedBy(-balanceData.currency.decimals).toNumber();
+                    let dataPoint = usdPricesEnabled ? balance.balanceUsd : tokenVolume;
+
+                    let point: IBalanceAreaChartPayload = {
+                        [CHART_DATA_KEY]: dataPoint,
+                        balanceToken: tokenVolume,
+                        timestamp: balance.timestamp
+                    };
+                    data.min = Math.min(data.min, dataPoint);
+                    data.max = Math.max(data.max, dataPoint);
+                    return point;
+                }).reverse();
+        } else {
+            if (usdPricesEnabled) {
+                // show total portfolio data, with USD prices
+                data.points = accountBalance.computeTotalBalance().map(balance => {
+                    let point: IBalanceAreaChartPayload = {
+                        [CHART_DATA_KEY]: balance.usd,
+                        balanceWei: balance.wei,
+                        timestamp: balance.timestamp
+                    };
+                    data.min = Math.min(data.min, balance.usd);
+                    data.max = Math.max(data.max, balance.usd);
+                    return point;
+                }).reverse();
+            } else {
+                // Show only ETH balance, because we can't aggregate all balances without USD prices
+                data.points = accountBalance.getEthBalance().chart.map(balance => {
+                    let dataPoint = weiToEth(balance.balance).toNumber();
+                    let point: IBalanceAreaChartPayload = {
+                        [CHART_DATA_KEY]: dataPoint,
+                        balanceWei: balance.balance,
+                        timestamp: balance.timestamp
+                    };
+                    data.min = Math.min(data.min, dataPoint);
+                    data.max = Math.max(data.max, dataPoint);
+                    return point;
+                }).reverse();
+            }
+        }
 
         // Ensure we don't get a "thin" chart stroke at the minimum
         if (data.max - data.min < 1) {