+++ src/app/eth-extended/module/account/balance/component/chart/IBalanceAreaChartPayload.ts
@@ -2,8 +2,11 @@ import { CHART_DATA_KEY } from "./BalanceChartData";
 import { BigNumber } from "app/util/BigNumber";
 
 export interface IBalanceAreaChartPayload {
-    /** Balance in USD */
+    /** Balance used to plot chart (USD, ETH or Token volume, depending on settings) */
     [CHART_DATA_KEY]: number;
-    balanceWei: BigNumber;
+    /** Balance in Wei (unset for token balances) */
+    balanceWei?: BigNumber;
+    /** Balance in token volume (for token balances) */
+    balanceToken?: number;
     timestamp: number;
 }